Check that BSS is in use
In commit 41e14bc6da05 ("Use libmnl instead of libnl"), I got rid of the
check for NL80211_BSS_STATUS because I assumed it wasn't necessary. It
turns out that netlink will return BSSs which are not in use (e.g., BSSs
which we found when doing an AP scan). NL80211_BSS_STATUS is only
returned if the BSS is used, so bring back the check.